Remove unecessary cast

+++ forwarder/cmd/forwarder/forwarder.go
@@ -168,7 +168,7 @@ func main() {
 		logger.Fatalf("error while running: %v", err)
 	}
 
-	timeMetric := MetricEntry{"Timespan", int64(time.Since(start).Milliseconds()), (time.Now()).Unix()}
+	timeMetric := MetricEntry{"Timespan", time.Since(start).Milliseconds(), (time.Now()).Unix()}
 	metrics := MetricValues{[]MetricEntry{timeMetric}}
 	metricBuffer, err := json.Marshal(metrics)
